requirements.txt
crewai
crewai[tools]
streamlit
python-dotenv
pyyaml
requests
tavily-python
litellm